source/requirements.txt
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.9.0
asttokens==3.0.0
certifi==2025.4.26
click==8.2.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
dnspython==2.7.0
executing==2.2.0
fonttools==4.58.1
flask
grpcio==1.72.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jedi==0.19.2
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.2.6
packaging==25.0
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.8
portalocker==2.10.1
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
pure_eval==0.2.3
pyahocorasick==2.1.0
pydantic_core==2.33.2
Pygments==2.19.1
pyngrok
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pywin32==310
pyzmq==26.4.0
qdrant-client==1.14.2
regex==2024.11.6
scikit-learn==1.6.1
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
textsearch==0.0.24
threadpoolctl==3.6.0
tornado==6.5.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
wcwidth==0.2.13
pymongo
python-dotenv
nltk
contractions
matplotlib
tqdm
scipy
pydantic
numpy